Update README.md
Data Analysis & Web Scraping
In this project, we will use web scraping techniques to collect data from an online furniture stores: Lluesma Muebles (https://www.muebleslluesma.com/). This web store sells Doos Design's and other competitors products. We will use the data to analyze how DOOS Design's products and prices compare to those of its rivals. We will also create visualizations to display the results of the analysis and to provide insights for DOOS Design to improve its market position and increase its sales.

Steps to follow:

Data was scraped from the website, using Python's Beautifulsoup library.
The scraped data was manipulated and cleaned using Python to create tables with relevant columns and values.
Power BI was used to create interactive charts and dashboards to discover useful patterns.
A report was created with clear and concise language to present the results and recommendations.
⚠️ For more extensive and in-depth analysis please refer to the report PDF.

🛠️ Python (Beautifulsoup / Pandas) and Power BI were used in this project.